requirements.txt
keras==2.2.4
tensorflow-gpu==1.14.0
fastapi
uvicorn
pillow==7.0.0
scikit-learn
imutils
matplotlib
